DAP: Fix windows support
- Added using named pipe for windows

+++ src/plugins/cmakeprojectmanager/fileapireader.cpp
@@ -98,9 +98,15 @@ void FileApiReader::parse(bool forceCMakeRun,
                                        + m_parameters.additionalCMakeArguments)
                                     : QStringList());
     if (debugging) {
-        FilePath file = FilePath::fromString(TemporaryDirectory::masterDirectoryPath() + "/cmake-dap.sock");
-        file.removeFile();
-        args << "--debugger" << "--debugger-pipe=" + file.path();
+        if (TemporaryDirectory::masterDirectoryFilePath().osType() == Utils::OsType::OsTypeWindows) {
+            args << "--debugger"
+                 << "--debugger-pipe \\\\.\\pipe\\cmake-dap";
+        } else {
+            FilePath file = TemporaryDirectory::masterDirectoryFilePath() / "cmake-dap.sock";
+            file.removeFile();
+            args << "--debugger"
+                 << "--debugger-pipe=" + file.path();
+        }
     }
 
     qCDebug(cmakeFileApiMode) << "Parameters request these CMake arguments:" << args;
